refactor: extract gigs app from blowcomotion monolith
