+++ blowcomotion/urls.py
@@ -8,6 +8,7 @@ from django.urls import include, path
 
 import blowcomotion.views as blowcomotion_views
 from blowcomotion import chart_api, member_urls
+from gigs import views as gigs_views
 from search import views as search_views
 
 urlpatterns = [
@@ -21,7 +22,7 @@ urlpatterns = [
     path("attendance/", blowcomotion_views.attendance_capture, name="attendance-main"),
     path("attendance/reports/", blowcomotion_views.attendance_reports, name="attendance-reports"),
     path("attendance/reports/<str:section_slug>/", blowcomotion_views.attendance_section_report_new, name="attendance-section-report"),
-    path("attendance/gigs-for-date/", blowcomotion_views.gigs_for_date, name="gigs-for-date"),
+    path("attendance/gigs-for-date/", gigs_views.gigs_for_date, name="gigs-for-date"),
     path("attendance/inactive-members/", blowcomotion_views.inactive_members, name="inactive-members"),
     path("attendance/<str:section_slug>/", blowcomotion_views.attendance_capture, name="attendance-capture"),
