+++ gigs/tests/test_sync_gigs_admin_view.py
